+++ src/util/enhancement-bonus-helper.mjs
@@ -1,7 +1,9 @@
-import { ammoEnhancementKey, ammoEnhancementStacksKey } from '../bonuses/ammunition-shared-keys.mjs';
+import { ammoBaneCreatureSubtype, ammoBaneCreatureType, ammoEnhancementKey, ammoEnhancementStacksKey } from '../bonuses/ammunition-shared-keys.mjs';
 import { MODULE_NAME } from '../consts.mjs';
 import { api } from './api.mjs';
+import { intersects } from './array-intersects.mjs';
 import { FormulaCacheHelper } from './flag-helpers.mjs';
+import { getIdsFromItem } from './get-id-array-from-flag.mjs';
 
 class EnhData {
     constructor({ base = 0, stacks = 0 }) {
@@ -17,6 +19,8 @@ class EnhBonusResult {
 
     /** @type {EnhData?} */
     action = null;
+
+    hasBane = false;
 
     get base() {
         return Math.max(this.ammo?.base || 0, this.action?.base || 0);
@@ -35,9 +39,10 @@ class EnhBonusResult {
  * @param {object} args
  * @param {ItemAction} [args.action]
  * @param {ItemLootPF} [args.ammo]
+ * @param {TokenPF[]} [args.targets]
  * @returns {EnhBonusResult}
  */
-const getEnhancementBonusForAction = ({ action, ammo }) => {
+const getEnhancementBonusForAction = ({ action, ammo, targets }) => {
     const enhData = new EnhBonusResult();
 
     if (action) {
@@ -57,7 +62,24 @@ const getEnhancementBonusForAction = ({ action, ammo }) => {
         }
 
         const base = FormulaCacheHelper.getModuleFlagValue(ammo, ammoEnhancementKey);
-        const stacks = FormulaCacheHelper.getModuleFlagValue(ammo, ammoEnhancementStacksKey);
+        let stacks = FormulaCacheHelper.getModuleFlagValue(ammo, ammoEnhancementStacksKey);
+
+        if (targets?.length) {
+            const creatureTypes = getIdsFromItem(ammo, ammoBaneCreatureType);
+            const creatureSubtypes = getIdsFromItem(ammo, ammoBaneCreatureSubtype);
+
+            if (creatureTypes.length || creatureSubtypes.length) {
+                enhData.hasBane = targets.map(x => x.actor).every((a) =>
+                    (!creatureTypes.length || intersects(creatureTypes, a?.race?.system.creatureTypes.total))
+                    && (!creatureSubtypes.length || intersects(creatureSubtypes, a?.race?.system.creatureSubtypes.total))
+                );
+
+                if (enhData.hasBane) {
+                    stacks += 2;
+                }
+            }
+        }
+
 
         enhData.ammo = new EnhData({ base, stacks });
     }